towards splinifying the basis
