+++ src/models/models.jl
@@ -10,9 +10,11 @@ include("radial_envelopes.jl")
 include("radial_transforms.jl")
 
 include("Rnl_basis.jl")
+include("Rnl_learnable.jl")
+include("Rnl_splines.jl")
+
 
 include("ace_heuristics.jl") 
-
 include("ace.jl")
 
 end
